requirements.txt
django
qrcode[pil]
python-dotenv
django-unfold
dj-database-url
psycopg2-binary
gunicorn
whitenoise
